requirements.txt
plumbum==1.8.2
pywin32==306
PyYAML==6.0.1
